Rename project to ansible-dev-environment
