+++ .gitignore
@@ -166,6 +166,6 @@ cython_debug/
 # and should all have detailed explanations
 
 # Version created and populated by setuptools_scm
-/src/ansible_development_environment/_version.py
+/src/ansible_dev_environment/_version.py
 
 .DS_Store
