+++ .vscode/launch.json
@@ -8,7 +8,7 @@
       "name": "Python: Module",
       "type": "python",
       "request": "launch",
-      "module": "ansible_development_environment",
+      "module": "ansible_dev_environment",
       "args": ["inspect", "-v", "-v", "-vv"],
       "justMyCode": true
     },
