+++ src/ansible_dev_environment/__main__.py
@@ -1,0 +1,11 @@
+"""A runpy entry point for ansible-dev-environment.
+
+This makes it possible to invoke CLI
+via :command:`python -m ansible_dev_environment`.
+"""
+
+from .cli import main
+
+
+if __name__ == "__main__":
+    main()
