+++ src/ansible_dev_environment/arg_parser.py
@@ -15,7 +15,7 @@ except ImportError:  # pragma: no cover
         import pkg_resources
 
         __version__ = pkg_resources.get_distribution(
-            "ansible_development_environment",
+            "ansible_dev_environment",
         ).version
     except Exception:  # pylint: disable=broad-except # noqa: BLE001
         # this is the fallback SemVer version picked by setuptools_scm when tag
@@ -33,7 +33,7 @@ def common_args(parser: ArgumentParser) -> None:
         "--lf",
         "--log-file <file>",
         dest="log_file",
-        default=str(Path.cwd() / "ansible-development-environment.log"),
+        default=str(Path.cwd() / "ansible-dev-environment.log"),
         help="Log file to write to.",
     )
     parser.add_argument(
