+++ src/ansible_dev_environment/config.py
@@ -1,4 +1,4 @@
-"""Constants, for now, for ansible-development-environment."""
+"""Constants, for now, for ansible-dev-environment."""
 
 from __future__ import annotations
 
@@ -51,7 +51,7 @@ class Config:
     @property
     def cache_dir(self: Config) -> Path:
         """Return the cache directory."""
-        cache_dir = self.venv / ".ansible-development-environment"
+        cache_dir = self.venv / ".ansible-dev-environment"
         if not cache_dir.exists():
             cache_dir.mkdir(parents=True)
         return cache_dir
