+++ src/ansible_dev_environment/subcommands/uninstaller.py
@@ -7,16 +7,16 @@ import shutil
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from ansible_development_environment.collection import (
+from ansible_dev_environment.collection import (
     Collection,
     parse_collection_request,
 )
-from ansible_development_environment.utils import collections_from_requirements
+from ansible_dev_environment.utils import collections_from_requirements
 
 
 if TYPE_CHECKING:
-    from ansible_development_environment.config import Config
-    from ansible_development_environment.output import Output
+    from ansible_dev_environment.config import Config
+    from ansible_dev_environment.output import Output
 
 
 class UnInstaller:
