--- src/ansible_development_environment/__main__.py
@@ -1,11 +1,0 @@
-"""A runpy entry point for ansible-development-environment.
-
-This makes it possible to invoke CLI
-via :command:`python -m ansible_development_environment`.
-"""
-
-from .cli import main
-
-
-if __name__ == "__main__":
-    main()
